refactor pick hero
