+++ game/scripts/vscripts/events.lua
@@ -60,18 +60,8 @@ function AIGameMode:OnGameStateChanged(keys)
             print("[AIGameMode] Setting pre-game options STRATEGY_TIME")
             self:PreGameOptions()
         end
-
-        Timers:CreateTimer(1, function()
-            for i = 0, (DOTA_MAX_TEAM_PLAYERS - 1) do
-                if PlayerResource:IsValidPlayer(i) then
-                    if PlayerResource:GetPlayer(i) and not PlayerResource:HasSelectedHero(i) then
-                        PlayerResource:GetPlayer(i):MakeRandomHeroSelection()
-                    end
-                end
-            end
-            self:EndScreenStats(1, false)
-        end)
     elseif state == DOTA_GAMERULES_STATE_PRE_GAME then
+        self:EndScreenStats(1, false)
         -- modifier towers
         local tTowers = Entities:FindAllByClassname("npc_dota_tower")
         local iTowerLevel = math.max(self.iGameDifficulty, 1)
